Added a new abstract PushEvent
